# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Retrieves information on Marvel characters",
        "title": "MarvelComicsAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/characters": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Characters"
                ],
                "summary": "Grt Ids of all characters",
                "operationId": "GetAllCharacters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/characters/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Characters"
                ],
                "summary": "Get a comic character by Id",
                "operationId": "GetCharacterById",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id of marvel character",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ComicCharacter"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ComicCharacter": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        }
    }
}